requirements.txt
python-dotenv
setuptools
pymupdf
PyPDF2
docx
python-docx
pywin32
pdf2image
langchain
langchain-community
langchain-ollama
langchain-openai
langchain-anthropic
langchain-google-genai
transformers
tiktoken
faiss-cpu
numpy
torch
torchvision
torchaudio
openai
rank_bm25
elasticsearch
haystack-ai
sentence-transformers
streamlit
